moved presence emitter into it's own module
